requirements.txt
absl-py==0.7.0
aniso8601==5.1.0
asn1crypto==0.24.0
astor==0.7.1
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
cryptography==2.5
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
future==0.17.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
line-bot-sdk==1.8.0
Markdown==3.0.1
MarkupSafe==1.1.1
numpy==1.16.1
pandas==0.24.1
paramiko==2.4.2
protobuf==3.6.1
pyasn1==0.4.5
pycparser==2.19
pymongo==3.7.2
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==4.2b1
relativetimebuilder==0.2.0
requests==2.21.0
scipy==1.2.1
scp==0.13.0
six==1.12.0
SQLAlchemy==1.2.18
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
